requirements.txt
#aerospike==2.1.3
amqp==1.4.9
anyjson==0.3.3
#apt-xapian-index==0.45
asn1crypto==0.24.0
Babel==2.5.0
backports-abc==0.5
backports.functools-lru-cache==1.6.4
backports.ssl-match-hostname==3.5.0.1
BeautifulSoup==3.2.2
beautifulsoup4==4.9.3
billiard==3.3.0.23
#bitarray==0.9.3
boto==2.48.0
boto3==1.4.7
botocore==1.7.7
CacheControl==0.12.5
cachetools==2.0.1
celery==3.1.25
certifi==2020.12.5
cffi==1.12.2
chardet==4.0.0
Cheetah==2.4.4
#cloud-init==0.7.5
cloudscraper==1.2.58
colorama==0.2.5
colorthief==0.2.1
configobj==4.7.2
contextlib2==0.6.0.post1
cryptography==2.7
decorator==4.4.2
defusedxml==0.7.1
dill==0.2.7.1
Django==1.8
django-adminplus==0.5
django-appconf==1.0.3
django-bootstrap-toolkit==2.15.0
django-celery==3.2.1
django-chunked-upload==2.0.0
django-common-helpers==0.9.1
django-compressor==2.4
django-cookies-samesite==0.9.0
django-cprofile-middleware==1.0.5
django-cron==0.5.0
django-endless-pagination==2.0
django-extensions==1.9.0
django-filter==1.0.4
#django-haystack==2.6.1
django-longerusernameandemail==0.5.7
django-messages==0.5.3
django-mobile==0.7.0
django-piston==0.2.3
django-redis==4.3.0
django-redisboard==3.0.2
django-rest-hooks==1.5.0
django-social-auth==0.7.28
django-socketio==0.3.8
django-storages==1.6.5
django-templated-email==0.4.9
django-templatetag-handlebars==1.3.1
django-user-agents==0.4.0
djangorestframework==3.6.4
docutils==0.14
elastic-apm==5.2.2
elasticsearch==5.4.0
elasticsearch-dsl==5.3.0
elementpath==1.3.3
enum34==1.1.6
favicon==0.7.0
#ffmpeg-python==0.2.0
firebase-admin==3.0.0
firebase-token-generator==2.0.1
flower==0.9.2
future==0.16.0
futures==3.3.0
gax-google-logging-v2==0.8.3
gax-google-pubsub-v1==0.8.3
gcloud==0.17.0
geopy==1.11.0
#gevent==0.13.6
gevent-socketio==0.2.1
gevent-websocket==0.2.3
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.0.2
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.5.0
google-cloud-storage==1.20.0
google-cloud-vision==0.39.0
google-gax==0.15.14
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
greenlet==0.4.2
grpc-google-logging-v2==0.8.1
grpc-google-pubsub-v1==0.8.1
grpcio==1.24.0
gunicorn==19.7.1
html5lib==0.999999999
httplib2==0.10.3
idna==2.10
importlib-resources==3.3.1
#inbloom==0.2.2
ipaddress==1.0.22
iso8601==0.1.14
jmespath==0.9.3
Js2Py==0.66
jsonpatch==1.3
jsonpath-rw==1.4.0
jsonpointer==1.0
jws==0.1.3
#kombu==3.0.37
#Landscape-Client==14.12
m3u8==0.6.0
Markdown==2.6.9
mechanize==0.3.5
meld3==0.6.10
#mmh3==2.5.1
msgpack==0.6.2
#MySQL-python==1.2.3
Naked==0.1.31
ndg-httpsclient==0.4.2
numpy==1.16.6
oauth==1.0.1
oauth2==1.9.0.post1
oauth2client==4.1.0
oauthlib==2.0.3
olefile==0.44
#opencv-python==4.1.1.26
#PAM==0.4.2
pathlib==1.0.1
pathlib2==2.3.6
pdf2image==1.10.0
#pexpect==3.1
pika==1.1.0
Pillow==4.2.1
ply==3.8
prettytable==0.7.2
protobuf==3.9.2
#py-bcrypt==0.4
pyasn1==0.3.3
pyasn1-modules==0.1.1
pycparser==2.19
#pycrypto==2.6.1
#pycryptodome==3.4.3
#pycurl==7.19.3
pyjsparser==2.7.1
PyJWT==1.5.3
pymongo==3.6.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
Pyrebase==3.0.27
pysaml2 @ git+https://github.com/masterapps-au/pysaml2.git@5cbbd447522a6238c10f9a0eec7e1e2bed553a0d
pyserial==2.6
#python-apt===0.9.3.5ubuntu2
python-dateutil==2.6.1
#python-debian===0.1.21-nmu2ubuntu2
python-firebase==1.2
python-jwt==2.0.2
python-memcached==1.58
python-openid==2.2.5
python-social-auth==0.2.14
pytz==2017.2
PyYAML==3.10
#rcssmin==1.0.6
redis==2.10.6
requests==2.25.1
requests-oauthlib==0.8.0
requests-toolbelt==0.9.1
#rjsmin==1.1.0
rsa==3.4.2
s3transfer==0.1.11
scalable-cuckoo-filter==1.1
scandir==1.10.0
shortuuid==0.5.0
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.15.0
slackclient==1.3.2
social-auth-app-django==3.4.0
social-auth-core==3.3.3
soupsieve==1.9.6
sphinx-me==0.3
ssh-import-id==3.21
StringGenerator==0.2.0
superlance==1.0.0
supervisor==3.0b2
tabulate==0.7.7
tldextract==1.7.1
tornado==4.5.2
#Twisted-Core==13.2.0
#Twisted-Names==13.2.0
#Twisted-Web==13.2.0
typing==3.10.0.0
tzlocal==1.5.1
ua-parser==0.10.0
uritemplate==3.0.0
urllib3==1.26.4
user-agents==2.2.0
virtualenv==16.7.5
webencodings==0.5.1
websocket-client==0.54.0
Whoosh==2.7.4
xmlschema==1.0.18
zipp==1.2.0
zope.interface==4.0.5